+++ js/src/utilities/settings.js
@@ -1,0 +1,1 @@
+export const getPwaSettings = () => (drupalSettings && drupalSettings.pwa_settings) || null;