+++ js/src/utilities/transform.course.js
@@ -6,6 +6,7 @@ import {
   transformCourseCategory,
   courseCategoryPropTypes,
 } from '@anu/utilities/transform.courseCategory';
+import { transformCoursesPage } from '@anu/utilities/transform.courses';
 
 /**
  * Transform course node from Drupal backend
@@ -17,8 +18,29 @@ const transformCourse = (node, data) => {
     return null;
   }
 
+  const courseId = fields.getNumberValue(node, 'nid');
+  // Prepare courses pages for current course.
+  let coursesPages = [];
+  if (data.courses_pages_by_course) {
+    data.courses_pages_by_course.map((set) => {
+      if (set.course_id == courseId) {
+        coursesPages = set.courses_pages;
+      }
+    });
+  }
+
+  // Prepare first lesson url for current course.
+  let firstLessonUrl = '';
+  if (data.first_lesson_url_by_course) {
+    data.first_lesson_url_by_course.map((set) => {
+      if (set.course_id == courseId) {
+        firstLessonUrl = set.first_lesson_url;
+      }
+    });
+  }
+
   return {
-    id: fields.getNumberValue(node, 'nid'),
+    id: courseId,
     title: fields.getTextValue(node, 'title'),
     description: fields.getTextValue(node, 'field_course_description'),
     url: fields.getNodeUrl(node),
@@ -37,6 +59,8 @@ const transformCourse = (node, data) => {
         .filter((lesson) => !!lesson),
       quiz: transformQuiz(fields.getArrayValue(module, 'field_module_assessment')[0], data),
     })),
+    courses_pages: coursesPages.map((coursesPage) => transformCoursesPage({ data: coursesPage })),
+    first_lesson_url: firstLessonUrl,
   };
 };
 
@@ -62,6 +86,8 @@ const coursePropTypes = PropTypes.shape({
     })
   ),
   labels: PropTypes.arrayOf(PropTypes.string),
+  courses_pages: PropTypes.arrayOf(PropTypes.shape({})),
+  first_lesson_url: PropTypes.string,
 });
 
 export { transformCourse, coursePropTypes };