+++ src/Controller/AnulmsNodeViewController.php
@@ -86,12 +86,14 @@ class AnulmsNodeViewController extends NodeViewController {
    *   The entity type manager.
    * @param \Drupal\anu_lms\Normalizer $normalizer
    *   The normalizer.
+   * @param \Drupal\anu_lms\CoursesPage $coursesPage
+   *   The Courses page service.
+   * @param \Drupal\anu_lms\Course $course
+   *   The Course service.
    * @param \Drupal\anu_lms\Lesson $lesson
    *   The Lesson service.
    * @param \Drupal\anu_lms\Quiz $quiz
    *   The Quiz service.
-   * @param \Drupal\anu_lms\CoursesPage $coursesPage
-   *   The Courses page service.
    * @param \Drupal\Core\Render\RendererInterface $renderer
    *   The renderer service.
    * @param \Drupal\Core\Session\AccountInterface $current_user
@@ -210,10 +212,6 @@ class AnulmsNodeViewController extends NodeViewController {
     // Attaches general site settings.
     $data['settings'] = $this->anulmsSettings->getSettings();
 
-    if (\Drupal::moduleHandler()->moduleExists('pwa')) {
-      $data['pwa'] = $this->anulmsSettings->getPwaSettings();
-    }
-
     // You can use `jQuery('#application').data('application')` in console for debug.
     $build['application'] = [
       '#type' => 'html_tag',
@@ -233,6 +231,11 @@ class AnulmsNodeViewController extends NodeViewController {
       $build['#attached'] = [
         'library' => ['anu_lms/lesson'],
       ];
+    }
+
+    // Attach PWA cache version.
+    if (\Drupal::moduleHandler()->moduleExists('pwa')) {
+      $build['#attached']['drupalSettings']['pwa_settings'] = $this->anulmsSettings->getPwaSettings();
     }
 
     // Disable cache for this page. @todo can be improved using cache tags.