+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+ - Adds offline mode support for using with Progressive Web App module
+ - Adds ability to translate strings for quiz submit buttons and info message
+ - Improve data migration script for update from 1.x
+ - Improve Course card on Courses page to lead directly to first lesson
+
 ## [2.2.0]
 
  - Added linear progress support for courses